nba-service/requirements.txt
nba_api==1.10.1
supabase==2.3.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pandas==2.1.4
requests==2.31.0
